move cask/cmd/info to cask/info
