+++ Library/Homebrew/cask/cmd.rb
@@ -12,7 +12,6 @@ require "cask/config"
 require "cask/cmd/abstract_command"
 require "cask/cmd/audit"
 require "cask/cmd/fetch"
-require "cask/cmd/info"
 require "cask/cmd/install"
 require "cask/cmd/list"
 require "cask/cmd/reinstall"
